refactor(editor): CommandItem → extensions/types; keep the command set in commandMenu
Move the shared CommandItem type into a new types-only extensions/types.ts,
and pull the slash-menu command definitions (COMMANDS + filterCommands) back
into commandMenu.ts beside the extension that consumes them.

components.tsx now keeps only the menu UI (CommandList) and references
CommandItem with a type-only import (erased at build), so it no longer imports
canUploadImages/promptImageUpload from extensions/images at runtime — removing
one of its components.tsx -> extensions/ edges.
