+++ frontend/src/lib/editor/components.tsx
@@ -2,7 +2,7 @@
 
 /** Tiptap-based live editor. Replaces `frontend/src/lib/editor/` (the
  * CodeMirror/OT-era editor, deleted once this cutover lands). */
-import { posToDOMRect, type Editor, type Range } from "@tiptap/core";
+import { posToDOMRect, type Editor } from "@tiptap/core";
 import { EditorContent, useEditor } from "@tiptap/react";
 import type {
   SuggestionKeyDownProps,
@@ -19,22 +19,8 @@ import {
 import { Awareness } from "y-protocols/awareness";
 import * as Y from "yjs";
 import { LineItemButton } from "@onyx-ai/opal/components";
-import {
-  SvgCheckSquare,
-  SvgCode,
-  SvgHash,
-  SvgImage,
-  SvgListTree,
-  SvgMinus,
-  SvgQuoteStart,
-  SvgTextLines,
-} from "@onyx-ai/opal/icons";
-import type { IconFunctionComponent } from "@onyx-ai/opal/types";
 import { tiptapExtensions } from "@/lib/editor/extensions";
-import {
-  canUploadImages,
-  promptImageUpload,
-} from "@/lib/editor/extensions/images";
+import type { CommandItem } from "@/lib/editor/extensions/types";
 import {
   commentHighlights as commentHighlightPlugin,
   sourceHighlights as sourceHighlightPlugin,
@@ -480,127 +466,6 @@ export function CoeditPresenceBar({
       ))}
     </div>
   );
-}
-
-export interface CommandItem {
-  title: string;
-  icon: IconFunctionComponent;
-  run: (editor: Editor, range: Range) => void;
-  /** Omitted means always offered. */
-  available?: (editor: Editor) => boolean;
-}
-
-const COMMANDS: CommandItem[] = [
-  {
-    title: "Text",
-    icon: SvgTextLines,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).setNode("paragraph").run(),
-  },
-  {
-    title: "Heading 1",
-    icon: SvgHash,
-    run: (editor, range) =>
-      editor
-        .chain()
-        .focus()
-        .deleteRange(range)
-        .setNode("heading", { level: 1 })
-        .run(),
-  },
-  {
-    title: "Heading 2",
-    icon: SvgHash,
-    run: (editor, range) =>
-      editor
-        .chain()
-        .focus()
-        .deleteRange(range)
-        .setNode("heading", { level: 2 })
-        .run(),
-  },
-  {
-    title: "Heading 3",
-    icon: SvgHash,
-    run: (editor, range) =>
-      editor
-        .chain()
-        .focus()
-        .deleteRange(range)
-        .setNode("heading", { level: 3 })
-        .run(),
-  },
-  {
-    title: "Bullet List",
-    icon: SvgListTree,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).toggleBulletList().run(),
-  },
-  {
-    title: "Numbered List",
-    icon: SvgListTree,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).toggleOrderedList().run(),
-  },
-  {
-    title: "Task List",
-    icon: SvgCheckSquare,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).toggleTaskList().run(),
-  },
-  {
-    title: "Blockquote",
-    icon: SvgQuoteStart,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).toggleBlockquote().run(),
-  },
-  {
-    title: "Code Block",
-    icon: SvgCode,
-    run: (editor, range) =>
-      editor.chain().focus().deleteRange(range).toggleCodeBlock().run(),
-  },
-  {
-    title: "Divider",
-    icon: SvgMinus,
-    run: (editor, range) => {
-      const node = editor.schema.nodes.thematic_break!.create(
-        // _raw: "1" matches what the backend stamps on every opaque block
-        // (see blocks.ts's own Enter-conversion path) — serialize_block's
-        // opaque-block fallback requires this exact attr, so a divider
-        // created without it fails every checkpoint from here on with
-        // NotImplementedError, permanently stranding edits in the update log.
-        { _raw: "1" },
-        editor.schema.text("---\n"),
-      );
-      editor
-        .chain()
-        .focus()
-        .deleteRange(range)
-        .insertContent(node.toJSON())
-        .run();
-    },
-  },
-  {
-    title: "Image",
-    icon: SvgImage,
-    // A view with no page path cannot upload, so the picker would discard the
-    // file it collected.
-    available: (editor) => canUploadImages(editor.view),
-    run: (editor, range) => {
-      // Close the menu first: the OS dialog steals focus, and the leftover
-      // "/image" text would otherwise survive in the doc behind it.
-      editor.chain().focus().deleteRange(range).run();
-      promptImageUpload(editor.view);
-    },
-  },
-];
-
-export function filterCommands(query: string, editor: Editor): CommandItem[] {
-  const available = COMMANDS.filter((c) => c.available?.(editor) ?? true);
-  const q = query.trim().toLowerCase();
-  if (!q) return available;
-  return available.filter((c) => c.title.toLowerCase().includes(q));
 }
 
 export interface CommandListHandle {
